move tests of commands to test_commands
